:wrench: Change default schema for test setup.
Why? To make use of the database schema restore in db/initializers/setup.sql.
I was originally hoping to move everything to the default dbo schema, but this seems like it will be too big a problem to solve for now.

Instead I'm opting to use a database restore, then truncate all tables for test.