:bug: Parse v-currency macro-ed input.
Why? v-currency returns the input as a string, but provides a parse function to convert to a number.

See https://vue-currency-input-v1.netlify.app/guide/#getting-the-number-value

NOTE: this is not a complete fix because the database column is an integer.
I'm going to investigate further, but I believe the best option would be to save everything in the database as a value in cents. e.g. book_supplies_cost_in_cents and then convert it in the UI.