:beetle: Fix student from draft conversion of high school id.
The front-end `student-finanical-aid` is setting this line of code
```
this.yukonHighSchools.unshift({ id: -1, name: "I did not attend a Yukon high school" });
```

But the -1 id does not map to anything in the back end. The intent appears to be to map to the special "Not Listed" record in the `high_school` table, so I'm mapping on creation.