:sparkles: Build out yukon universtity and alkan air templates.
There is a lot of duplication between these templates, but I felt that it would be a good idea to make it easy for them to vary independently at this time.

At some future date we might want to build some partials and shared serialization patterns, but for a first pass, its probably safest to keep everything separate.